requirements.txt
pyserial
matplotlib
numpy
pyaudio
